% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parn88-package.R
\docType{package}
\name{parn88-package}
\alias{parn88}
\alias{parn88-package}
\title{parn88: Implementation of Parncutt's (1988) Root-Finding Algorithm}
\description{
This package provides an implementation of Parncutt's (1988) root-finding model, as described in the Music Perception paper entitled "Revision of Terhardt's psychoacoustical model of the root(s) of a musical chord".
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/pmcharrison/parn88}
  \item Report bugs at \url{https://github.com/pmcharrison/parn88/issues}
}

}
\author{
\strong{Maintainer}: Peter Harrison \email{pmc.harrison@gmail.com}

}
\keyword{internal}
